修改战棋部分架构
