+++ THE_Client/Assets/BattleScene/Map.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 91a3cb0ac86e9a54da5a802e0ceb8608
+folderAsset: yes
+timeCreated: 1456152504
+licenseType: Pro
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
